+++ client/src/App.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { MuiThemeProvider } from "@material-ui/core";
-import { BrowserRouter, Switch, Route } from "react-router-dom";
+import { BrowserRouter, Switch } from "react-router-dom";
 
 import { theme } from "./themes/theme";
 import Profile from "./pages/Profile";
@@ -15,8 +15,6 @@ import AuthProvider from "./context/AuthContext";
 import OnboardingPage from "./pages/OnboardingPage";
 import "./App.css";
 
-
-
 function App() {
 	return (
 		<MuiThemeProvider theme={theme}>
@@ -30,7 +28,6 @@ function App() {
 						<PrivateRoute path="/checkout" component={Checkout} />
 						<PrivateRoute path="/reviews" exact component={Reviews} />
 						<PrivateRoute path="/onboarding" exact component={OnboardingPage} />
-						
 					</Switch>
 				</BrowserRouter>
 			</AuthProvider>
