+++ client/src/pages/OnboardingPage.js
@@ -2,12 +2,11 @@ import React from "react";
 import WithBackground from "../hocs/withBackground";
 
 import Onboarding from "../components/Onboarding";
-import NewOnboard from "../components/NewOnboarding";
 
 const OnboardingPage = () => {
 	return (
 		<WithBackground>
-			<NewOnboard />
+			<Onboarding />
 		</WithBackground>
 	);
 };
