+++ client/src/prism.css
@@ -51,7 +51,7 @@
 }
 
 .RichEditor-editor .public-DraftStyleDefault-pre {
-  background-color: rgba(0, 0, 0, 0.9);
+  background-color: rgba(0, 0, 0, 0.85);
   font-family: "Inconsolata", "Menlo", "Consolas", monospace;
   font-size: 16px;
 	padding: 20px;
